Moved localPlayer.tsx outside pages
