+++ src/pages/mainpage.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react';
-import { setPlayerName, myself } from './localPlayer';
+import { setPlayerName, myself } from '../localPlayer';
 
 type MainPageProps = {
   socket: any;
